pyproject.toml
[project]
name = "reddit-intel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=0.30.0",
    "praw>=7.8.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
